// src/enums/analyze-strategies.enum.ts
export enum AnalyzeStrategy {
  RSI_DIVERGENCE = "RSI_DIVERGENCE",
  RSI_TRIPLE_DIVERGENCE = "RSI_TRIPLE_DIVERGENCE",
  RSI_WITH_BB = "RSI_WITH_BB",
  NOT_AVAILABLE = "NOT_AVAILABLE",
  PUMP_OR_DUMP = "PUMP_OR_DUMP",
  SWING_HIGH_LOW = "SWING_HIGH_LOW",
  RSI_3_DIVERGENCE = "RSI_3_DIVERGENCE",
  HIGHEST_BB = "HIGHEST_BB",
  LOWEST_BB = "LOWEST_BB",
  HIGHEST_RSI = "HIGHEST_RSI",
  LOWEST_RSI = "LOWEST_RSI",
  RSI_3PERIOD_OVEREXTEND = "RSI_3PERIOD_OVEREXTEND",
}